refine tests and simplify logic
