+++ test/e2e/app-dir/reexport-client-component-metadata/app/(group)/layout.tsx
@@ -1,5 +1,8 @@
-'use client'
-
 export default function Layout({ children }: { children: React.ReactNode }) {
   return <div>{children}</div>
+}
+
+export const metadata = {
+  title: 'Root Layout',
+  description: 'Root Description',
 }
