+++ test/e2e/app-dir/reexport-client-component-metadata/app/(group)/no-override/page-content.tsx
@@ -1,0 +1,5 @@
+'use client'
+
+export default function PageContent() {
+  return <h1>Page 2 Content</h1>
+}
