--- test/e2e/app-dir/root-layout-metadata/app/[lang]/(group)/two/page-content.tsx
@@ -1,14 +1,0 @@
-'use client'
-
-import { useEffect } from 'react'
-
-export default function PageContent() {
-  useEffect(() => {
-    console.log('Page 1 Content Mounted')
-    return () => {
-      console.log('Page 1 Content Unmounted')
-    }
-  }, [])
-
-  return <h1>Page 2 Content</h1>
-}
